moved listen call out of server/index so multiple test files can run
